Refactor, and rename
- Also moves to be closed to the db context its initializing
